+++ src/components/link-preview/soundcloud.jsx
@@ -68,7 +68,7 @@ export default function SoundCloudPreview({ url }) {
         style={{ width: '100%', maxWidth: '450px', height: '166px' }}
       />
       <div className="info">
-        <a href={url} target="_blank" title={byline}>
+        <a href={url} target="_blank" title={byline} rel="noreferrer">
           {byline} at SoundCloud
         </a>
       </div>