+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.141.0] - Not released
+### Changed
+- Migrate to React 19.
+- Migrate to ESLint 9 with flat config.
+- Split giant middlewares file into smaller files.
 
 ## [1.140.2] - 2025-09-20
 ### Fixed